Usually there's no need to delete surcharges because they are dynamic

+++ system/modules/isotope/library/Isotope/Model/ProductCollection.php
@@ -509,7 +509,9 @@ abstract class ProductCollection extends TypeAgent
         }
 
         foreach ($this->getSurcharges() as $objSurcharge) {
-            $objSurcharge->delete();
+            if ($objSurcharge->id) {
+                $objSurcharge->delete();
+            }
         }
 
         $this->arrItems = null;
